Struct aws_sdk_appstream::client::fluent_builders::CreateStack

pub struct CreateStack<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> CreateStack<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(self) -> Result<CreateStackOutput, SdkError<CreateStackError>> where
R::Policy: SmithyRetryPolicy<CreateStackInputOperationOutputAlias, CreateStackOutput, CreateStackError, CreateStackInputOperationRetryAlias>,

pub fn name(self, inp: impl Into<String>) -> Self

The name of the stack.

pub fn set_name(self, input: Option<String>) -> Self

pub fn description(self, inp: impl Into<String>) -> Self

The description to display.

pub fn set_description(self, input: Option<String>) -> Self

pub fn display_name(self, inp: impl Into<String>) -> Self

The stack name to display.

pub fn set_display_name(self, input: Option<String>) -> Self

pub fn storage_connectors(self, inp: impl Into<StorageConnector>) -> Self

Appends an item to StorageConnectors.

To override the contents of this collection use set_storage_connectors.

The storage connectors to enable.

pub fn set_storage_connectors(
self,
input: Option<Vec<StorageConnector>>
) -> Self

pub fn redirect_url(self, inp: impl Into<String>) -> Self

The URL that users are redirected to after their streaming session ends.

pub fn set_redirect_url(self, input: Option<String>) -> Self

pub fn feedback_url(self, inp: impl Into<String>) -> Self

The URL that users are redirected to after they click the Send Feedback link. If no URL is specified, no Send Feedback link is displayed.

pub fn set_feedback_url(self, input: Option<String>) -> Self

pub fn user_settings(self, inp: impl Into<UserSetting>) -> Self

Appends an item to UserSettings.

To override the contents of this collection use set_user_settings.

The actions that are enabled or disabled for users during their streaming sessions. By default, these actions are enabled.

pub fn set_user_settings(self, input: Option<Vec<UserSetting>>) -> Self

pub fn application_settings(self, inp: ApplicationSettings) -> Self

The persistent application settings for users of a stack. When these settings are enabled, changes that users make to applications and Windows settings are automatically saved after each session and applied to the next session.

pub fn set_application_settings(
self,
input: Option<ApplicationSettings>
) -> Self

pub fn tags(self, k: impl Into<String>, v: impl Into<String>) -> Self

Adds a key-value pair to Tags.

To override the contents of this collection use set_tags.

The tags to associate with the stack. A tag is a key-value pair, and the value is optional. For example, Environment=Test. If you do not specify a value, Environment=.

If you do not specify a value, the value is set to an empty string.

Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following special characters:

_ . : / = + \ - @

For more information about tags, see Tagging Your Resources in the Amazon AppStream 2.0 Administration Guide.

pub fn set_tags(self, input: Option<HashMap<String, String>>) -> Self

pub fn access_endpoints(self, inp: impl Into<AccessEndpoint>) -> Self

Appends an item to AccessEndpoints.

To override the contents of this collection use set_access_endpoints.

The list of interface VPC endpoint (interface endpoint) objects. Users of the stack can connect to AppStream 2.0 only through the specified endpoints.

pub fn set_access_endpoints(self, input: Option<Vec<AccessEndpoint>>) -> Self

pub fn embed_host_domains(self, inp: impl Into<String>) -> Self

Appends an item to EmbedHostDomains.

To override the contents of this collection use set_embed_host_domains.

The domains where AppStream 2.0 streaming sessions can be embedded in an iframe. You must approve the domains that you want to host embedded AppStream 2.0 streaming sessions.

pub fn set_embed_host_domains(self, input: Option<Vec<String>>) -> Self