Module aws_sdk_appstream::error

Modules

concurrent_modification_exception

See ConcurrentModificationException

incompatible_image_exception

See IncompatibleImageException

invalid_account_status_exception

See InvalidAccountStatusException

invalid_parameter_combination_exception

See InvalidParameterCombinationException

invalid_role_exception

See InvalidRoleException

limit_exceeded_exception

See LimitExceededException

operation_not_permitted_exception

See OperationNotPermittedException

request_limit_exceeded_exception

See RequestLimitExceededException

resource_already_exists_exception

See ResourceAlreadyExistsException

resource_in_use_exception

See ResourceInUseException

resource_not_available_exception

See ResourceNotAvailableException

resource_not_found_exception

See ResourceNotFoundException

Structs

AssociateFleetError

BatchAssociateUserStackError

BatchDisassociateUserStackError

ConcurrentModificationException

An API error occurred. Wait a few minutes and try again.

CopyImageError

CreateDirectoryConfigError

CreateFleetError

CreateImageBuilderError

CreateImageBuilderStreamingURLError

CreateStackError

CreateStreamingURLError

CreateUpdatedImageError

CreateUsageReportSubscriptionError

CreateUserError

DeleteDirectoryConfigError

DeleteFleetError

DeleteImageBuilderError

DeleteImageError

DeleteImagePermissionsError

DeleteStackError

DeleteUsageReportSubscriptionError

DeleteUserError

DescribeDirectoryConfigsError

DescribeFleetsError

DescribeImageBuildersError

DescribeImagePermissionsError

DescribeImagesError

DescribeSessionsError

DescribeStacksError

DescribeUsageReportSubscriptionsError

DescribeUserStackAssociationsError

DescribeUsersError

DisableUserError

DisassociateFleetError

EnableUserError

ExpireSessionError

IncompatibleImageException

The image can't be updated because it's not compatible for updates.

InvalidAccountStatusException

The resource cannot be created because your AWS account is suspended. For assistance, contact AWS Support.

InvalidParameterCombinationException

Indicates an incorrect combination of parameters, or a missing parameter.

InvalidRoleException

The specified role is invalid.

LimitExceededException

The requested limit exceeds the permitted limit for an account.

ListAssociatedFleetsError

ListAssociatedStacksError

ListTagsForResourceError

OperationNotPermittedException

The attempted operation is not permitted.

RequestLimitExceededException

AppStream 2.0 can’t process the request right now because the Describe calls from your AWS account are being throttled by Amazon EC2. Try again later.

ResourceAlreadyExistsException

The specified resource already exists.

ResourceInUseException

The specified resource is in use.

ResourceNotAvailableException

The specified resource exists and is not in use, but isn't available.

ResourceNotFoundException

The specified resource was not found.

StartFleetError

StartImageBuilderError

StopFleetError

StopImageBuilderError

TagResourceError

UntagResourceError

UpdateDirectoryConfigError

UpdateFleetError

UpdateImagePermissionsError

UpdateStackError

Enums

AssociateFleetErrorKind

BatchAssociateUserStackErrorKind

BatchDisassociateUserStackErrorKind

CopyImageErrorKind

CreateDirectoryConfigErrorKind

CreateFleetErrorKind

CreateImageBuilderErrorKind

CreateImageBuilderStreamingURLErrorKind

CreateStackErrorKind

CreateStreamingURLErrorKind

CreateUpdatedImageErrorKind

CreateUsageReportSubscriptionErrorKind

CreateUserErrorKind

DeleteDirectoryConfigErrorKind

DeleteFleetErrorKind

DeleteImageBuilderErrorKind

DeleteImageErrorKind

DeleteImagePermissionsErrorKind

DeleteStackErrorKind

DeleteUsageReportSubscriptionErrorKind

DeleteUserErrorKind

DescribeDirectoryConfigsErrorKind

DescribeFleetsErrorKind

DescribeImageBuildersErrorKind

DescribeImagePermissionsErrorKind

DescribeImagesErrorKind

DescribeSessionsErrorKind

DescribeStacksErrorKind

DescribeUsageReportSubscriptionsErrorKind

DescribeUserStackAssociationsErrorKind

DescribeUsersErrorKind

DisableUserErrorKind

DisassociateFleetErrorKind

EnableUserErrorKind

ExpireSessionErrorKind

ListAssociatedFleetsErrorKind

ListAssociatedStacksErrorKind

ListTagsForResourceErrorKind

StartFleetErrorKind

StartImageBuilderErrorKind

StopFleetErrorKind

StopImageBuilderErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateDirectoryConfigErrorKind

UpdateFleetErrorKind

UpdateImagePermissionsErrorKind

UpdateStackErrorKind