Struct aws_sdk_appstream::input::update_fleet_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for UpdateFleetInput

Implementations

impl Builder

pub fn image_name(self, input: impl Into<String>) -> Self

The name of the image used to create the fleet.

pub fn set_image_name(self, input: Option<String>) -> Self

pub fn image_arn(self, input: impl Into<String>) -> Self

The ARN of the public, private, or shared image to use.

pub fn set_image_arn(self, input: Option<String>) -> Self

pub fn name(self, input: impl Into<String>) -> Self

A unique name for the fleet.

pub fn set_name(self, input: Option<String>) -> Self

pub fn instance_type(self, input: impl Into<String>) -> Self

The instance type to use when launching fleet instances. The following instance types are available:

stream.standard.small

stream.standard.medium

stream.standard.large

stream.compute.large

stream.compute.xlarge

stream.compute.2xlarge

stream.compute.4xlarge

stream.compute.8xlarge

stream.memory.large

stream.memory.xlarge

stream.memory.2xlarge

stream.memory.4xlarge

stream.memory.8xlarge

stream.memory.z1d.large

stream.memory.z1d.xlarge

stream.memory.z1d.2xlarge

stream.memory.z1d.3xlarge

stream.memory.z1d.6xlarge

stream.memory.z1d.12xlarge

stream.graphics-design.large

stream.graphics-design.xlarge

stream.graphics-design.2xlarge

stream.graphics-design.4xlarge

stream.graphics-desktop.2xlarge

stream.graphics.g4dn.xlarge

stream.graphics.g4dn.2xlarge

stream.graphics.g4dn.4xlarge

stream.graphics.g4dn.8xlarge

stream.graphics.g4dn.12xlarge

stream.graphics.g4dn.16xlarge

stream.graphics-pro.4xlarge

stream.graphics-pro.8xlarge

stream.graphics-pro.16xlarge

pub fn set_instance_type(self, input: Option<String>) -> Self

pub fn compute_capacity(self, input: ComputeCapacity) -> Self

The desired capacity for the fleet.

pub fn set_compute_capacity(self, input: Option<ComputeCapacity>) -> Self

pub fn vpc_config(self, input: VpcConfig) -> Self

The VPC configuration for the fleet.

pub fn set_vpc_config(self, input: Option<VpcConfig>) -> Self

pub fn max_user_duration_in_seconds(self, input: i32) -> Self

The maximum amount of time that a streaming session can remain active, in seconds. If users are still connected to a streaming instance five minutes before this limit is reached, they are prompted to save any open documents before being disconnected. After this time elapses, the instance is terminated and replaced by a new instance.

Specify a value between 600 and 360000.

pub fn set_max_user_duration_in_seconds(self, input: Option<i32>) -> Self

pub fn disconnect_timeout_in_seconds(self, input: i32) -> Self

The amount of time that a streaming session remains active after users disconnect. If users try to reconnect to the streaming session after a disconnection or network interruption within this time interval, they are connected to their previous session. Otherwise, they are connected to a new session with a new streaming instance.

Specify a value between 60 and 360000.

pub fn set_disconnect_timeout_in_seconds(self, input: Option<i32>) -> Self

pub fn delete_vpc_config(self, input: bool) -> Self

Deletes the VPC association for the specified fleet.

pub fn set_delete_vpc_config(self, input: Option<bool>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

The description to display.

pub fn set_description(self, input: Option<String>) -> Self

pub fn display_name(self, input: impl Into<String>) -> Self

The fleet name to display.

pub fn set_display_name(self, input: Option<String>) -> Self

pub fn enable_default_internet_access(self, input: bool) -> Self

Enables or disables default internet access for the fleet.

pub fn set_enable_default_internet_access(self, input: Option<bool>) -> Self

pub fn domain_join_info(self, input: DomainJoinInfo) -> Self

The name of the directory and organizational unit (OU) to use to join the fleet to a Microsoft Active Directory domain.

pub fn set_domain_join_info(self, input: Option<DomainJoinInfo>) -> Self

pub fn idle_disconnect_timeout_in_seconds(self, input: i32) -> Self

The amount of time that users can be idle (inactive) before they are disconnected from their streaming session and the DisconnectTimeoutInSeconds time interval begins. Users are notified before they are disconnected due to inactivity. If users try to reconnect to the streaming session before the time interval specified in DisconnectTimeoutInSeconds elapses, they are connected to their previous session. Users are considered idle when they stop providing keyboard or mouse input during their streaming session. File uploads and downloads, audio in, audio out, and pixels changing do not qualify as user activity. If users continue to be idle after the time interval in IdleDisconnectTimeoutInSeconds elapses, they are disconnected.

To prevent users from being disconnected due to inactivity, specify a value of 0. Otherwise, specify a value between 60 and 3600. The default value is 0.

If you enable this feature, we recommend that you specify a value that corresponds exactly to a whole number of minutes (for example, 60, 120, and 180). If you don't do this, the value is rounded to the nearest minute. For example, if you specify a value of 70, users are disconnected after 1 minute of inactivity. If you specify a value that is at the midpoint between two different minutes, the value is rounded up. For example, if you specify a value of 90, users are disconnected after 2 minutes of inactivity.

pub fn set_idle_disconnect_timeout_in_seconds(self, input: Option<i32>) -> Self

pub fn attributes_to_delete(self, input: impl Into<FleetAttribute>) -> Self

pub fn set_attributes_to_delete(
self,
input: Option<Vec<FleetAttribute>>
) -> Self

pub fn iam_role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the IAM role to apply to the fleet. To assume a role, a fleet instance calls the AWS Security Token Service (STS) AssumeRole API operation and passes the ARN of the role to use. The operation creates a new session with temporary credentials. AppStream 2.0 retrieves the temporary credentials and creates the appstream_machine_role credential profile on the instance.

For more information, see Using an IAM Role to Grant Permissions to Applications and Scripts Running on AppStream 2.0 Streaming Instances in the Amazon AppStream 2.0 Administration Guide.

pub fn set_iam_role_arn(self, input: Option<String>) -> Self

pub fn stream_view(self, input: StreamView) -> Self

The AppStream 2.0 view that is displayed to your users when they stream from the fleet. When APP is specified, only the windows of applications opened by users display. When DESKTOP is specified, the standard desktop that is provided by the operating system displays.

The default value is APP.

pub fn set_stream_view(self, input: Option<StreamView>) -> Self

pub fn build(self) -> Result<UpdateFleetInput, BuildError>

Consumes the builder and constructs a UpdateFleetInput