Module aws_sdk_appstream::model

Modules

access_endpoint

See AccessEndpoint

application

See Application

application_settings

See ApplicationSettings

application_settings_response

See ApplicationSettingsResponse

compute_capacity

See ComputeCapacity

compute_capacity_status

See ComputeCapacityStatus

directory_config

See DirectoryConfig

domain_join_info

See DomainJoinInfo

fleet

See Fleet

fleet_error

See FleetError

image

See Image

image_builder

See ImageBuilder

image_builder_state_change_reason

See ImageBuilderStateChangeReason

image_permissions

See ImagePermissions

image_state_change_reason

See ImageStateChangeReason

last_report_generation_execution_error

See LastReportGenerationExecutionError

network_access_configuration

See NetworkAccessConfiguration

resource_error

See ResourceError

service_account_credentials

See ServiceAccountCredentials

session

See Session

shared_image_permissions

See SharedImagePermissions

stack

See Stack

stack_error

See StackError

storage_connector

See StorageConnector

usage_report_subscription

See UsageReportSubscription

user

See User

user_setting

See UserSetting

user_stack_association

See UserStackAssociation

user_stack_association_error

See UserStackAssociationError

vpc_config

See VpcConfig

Structs

AccessEndpoint

Describes an interface VPC endpoint (interface endpoint) that lets you create a private connection between the virtual private cloud (VPC) that you specify and AppStream 2.0. When you specify an interface endpoint for a stack, users of the stack can connect to AppStream 2.0 only through that endpoint. When you specify an interface endpoint for an image builder, administrators can connect to the image builder only through that endpoint.

Application

Describes an application in the application catalog.

ApplicationSettings

The persistent application settings for users of a stack.

ApplicationSettingsResponse

Describes the persistent application settings for users of a stack.

ComputeCapacity

Describes the capacity for a fleet.

ComputeCapacityStatus

Describes the capacity status for a fleet.

DirectoryConfig

Describes the configuration information required to join fleets and image builders to Microsoft Active Directory domains.

DomainJoinInfo

Describes the configuration information required to join fleets and image builders to Microsoft Active Directory domains.

Fleet

Describes a fleet.

FleetError

Describes a fleet error.

Image

Describes an image.

ImageBuilder

Describes a virtual machine that is used to create an image.

ImageBuilderStateChangeReason

Describes the reason why the last image builder state change occurred.

ImagePermissions

Describes the permissions for an image.

ImageStateChangeReason

Describes the reason why the last image state change occurred.

LastReportGenerationExecutionError

Describes the error that is returned when a usage report can't be generated.

NetworkAccessConfiguration

Describes the network details of the fleet or image builder instance.

ResourceError

Describes a resource error.

ServiceAccountCredentials

Describes the credentials for the service account used by the fleet or image builder to connect to the directory.

Session

Describes a streaming session.

SharedImagePermissions

Describes the permissions that are available to the specified AWS account for a shared image.

Stack

Describes a stack.

StackError

Describes a stack error.

StorageConnector

Describes a connector that enables persistent storage for users.

UsageReportSubscription

Describes information about the usage report subscription.

User

Describes a user in the user pool.

UserSetting

Describes an action and whether the action is enabled or disabled for users during their streaming sessions.

UserStackAssociation

Describes a user in the user pool and the associated stack.

UserStackAssociationError

Describes the error that is returned when a user can’t be associated with or disassociated from a stack.

VpcConfig

Describes VPC configuration information for fleets and image builders.

Enums

AccessEndpointType

Action

AuthenticationType

FleetAttribute

The fleet attribute.

FleetErrorCode

FleetState

FleetType

ImageBuilderState

ImageBuilderStateChangeReasonCode

ImageState

ImageStateChangeReasonCode

MessageAction

Permission

PlatformType

SessionConnectionState

SessionState

Possible values for the state of a streaming session.

StackAttribute

StackErrorCode

StorageConnectorType

The type of storage connector.

StreamView

UsageReportExecutionErrorCode

UsageReportSchedule

UserStackAssociationErrorCode

VisibilityType