Module aws_sdk_appstream::operation

Structs

AssociateFleet

Associates the specified fleet with the specified stack.

BatchAssociateUserStack

Associates the specified users with the specified stacks. Users in a user pool cannot be assigned to stacks with fleets that are joined to an Active Directory domain.

BatchDisassociateUserStack

Disassociates the specified users from the specified stacks.

CopyImage

Copies the image within the same region or to a new region within the same AWS account. Note that any tags you added to the image will not be copied.

CreateDirectoryConfig

Creates a Directory Config object in AppStream 2.0. This object includes the configuration information required to join fleets and image builders to Microsoft Active Directory domains.

CreateFleet

Creates a fleet. A fleet consists of streaming instances that run a specified image.

CreateImageBuilder

Creates an image builder. An image builder is a virtual machine that is used to create an image.

CreateImageBuilderStreamingURL

Creates a URL to start an image builder streaming session.

CreateStack

Creates a stack to start streaming applications to users. A stack consists of an associated fleet, user access policies, and storage configurations.

CreateStreamingURL

Creates a temporary URL to start an AppStream 2.0 streaming session for the specified user. A streaming URL enables application streaming to be tested without user setup.

CreateUpdatedImage

Creates a new image with the latest Windows operating system updates, driver updates, and AppStream 2.0 agent software.

CreateUsageReportSubscription

Creates a usage report subscription. Usage reports are generated daily.

CreateUser

Creates a new user in the user pool.

DeleteDirectoryConfig

Deletes the specified Directory Config object from AppStream 2.0. This object includes the information required to join streaming instances to an Active Directory domain.

DeleteFleet

Deletes the specified fleet.

DeleteImage

Deletes the specified image. You cannot delete an image when it is in use.

DeleteImageBuilder

Deletes the specified image builder and releases the capacity.

DeleteImagePermissions

Deletes permissions for the specified private image. After you delete permissions for an image, AWS accounts to which you previously granted these permissions can no longer use the image.

DeleteStack

Deletes the specified stack. After the stack is deleted, the application streaming environment provided by the stack is no longer available to users. Also, any reservations made for application streaming sessions for the stack are released.

DeleteUsageReportSubscription

Disables usage report generation.

DeleteUser

Deletes a user from the user pool.

DescribeDirectoryConfigs

Retrieves a list that describes one or more specified Directory Config objects for AppStream 2.0, if the names for these objects are provided. Otherwise, all Directory Config objects in the account are described. These objects include the configuration information required to join fleets and image builders to Microsoft Active Directory domains.

DescribeFleets

Retrieves a list that describes one or more specified fleets, if the fleet names are provided. Otherwise, all fleets in the account are described.

DescribeImageBuilders

Retrieves a list that describes one or more specified image builders, if the image builder names are provided. Otherwise, all image builders in the account are described.

DescribeImagePermissions

Retrieves a list that describes the permissions for shared AWS account IDs on a private image that you own.

DescribeImages

Retrieves a list that describes one or more specified images, if the image names or image ARNs are provided. Otherwise, all images in the account are described.

DescribeSessions

Retrieves a list that describes the streaming sessions for a specified stack and fleet. If a UserId is provided for the stack and fleet,

DescribeStacks

Retrieves a list that describes one or more specified stacks, if the stack names are provided. Otherwise, all stacks in the account are described.

DescribeUsageReportSubscriptions

Retrieves a list that describes one or more usage report subscriptions.

DescribeUserStackAssociations

Retrieves a list that describes the UserStackAssociation objects. You must specify either or both of the following:

DescribeUsers

Retrieves a list that describes one or more specified users in the user pool.

DisableUser

Disables the specified user in the user pool. Users can't sign in to AppStream 2.0 until they are re-enabled. This action does not delete the user.

DisassociateFleet

Disassociates the specified fleet from the specified stack.

EnableUser

Enables a user in the user pool. After being enabled, users can sign in to AppStream 2.0 and open applications from the stacks to which they are assigned.

ExpireSession

Immediately stops the specified streaming session.

ListAssociatedFleets

Retrieves the name of the fleet that is associated with the specified stack.

ListAssociatedStacks

Retrieves the name of the stack with which the specified fleet is associated.

ListTagsForResource

Retrieves a list of all tags for the specified AppStream 2.0 resource. You can tag AppStream 2.0 image builders, images, fleets, and stacks.

StartFleet

Starts the specified fleet.

StartImageBuilder

Starts the specified image builder.

StopFleet

Stops the specified fleet.

StopImageBuilder

Stops the specified image builder.

TagResource

Adds or overwrites one or more tags for the specified AppStream 2.0 resource. You can tag AppStream 2.0 image builders, images, fleets, and stacks.

UntagResource

Disassociates one or more specified tags from the specified AppStream 2.0 resource.

UpdateDirectoryConfig

Updates the specified Directory Config object in AppStream 2.0. This object includes the configuration information required to join fleets and image builders to Microsoft Active Directory domains.

UpdateFleet

Updates the specified fleet.

UpdateImagePermissions

Adds or updates permissions for the specified private image.

UpdateStack

Updates the specified fields for the specified stack.